+++ source/world.c
@@ -251,7 +251,7 @@ areanode_t *SV_CreateAreaNode (int depth, vec3_t mins, vec3_t maxs)
 	else
 		anode->axis = 1;
 
-	anode->dist = 0.5 * (maxs[anode->axis] + mins[anode->axis]);
+	anode->dist = 0.5f * (maxs[anode->axis] + mins[anode->axis]);
 	VectorCopy (mins, mins1);
 	VectorCopy (mins, mins2);
 	VectorCopy (maxs, maxs1);
@@ -696,8 +696,8 @@ reenter:
 	}
 	else
 	{
-		t1 = DoublePrecisionDotProduct (plane->normal, p1) - plane->dist;
-		t2 = DoublePrecisionDotProduct (plane->normal, p2) - plane->dist;
+		t1 = DoublePrecisionDotProduct (plane->normal, p1) - (double)plane->dist;
+		t2 = DoublePrecisionDotProduct (plane->normal, p2) - (double)plane->dist;
 	}
 
 	/*if its completely on one side, resume on that side*/
@@ -744,19 +744,19 @@ reenter:
 		/*we impacted the back of the node, so flip the plane*/
 		trace->plane.dist = -plane->dist;
 		VectorNegate(plane->normal, trace->plane.normal);
-		midf = (t1 + DIST_EPSILON) / (t1 - t2);
+		midf = (t1 + (float)DIST_EPSILON) / (t1 - t2);
 	}
 	else
 	{
 		/*we impacted the front of the node*/
 		trace->plane.dist = plane->dist;
 		VectorCopy(plane->normal, trace->plane.normal);
-		midf = (t1 - DIST_EPSILON) / (t1 - t2);
+		midf = (t1 - (float)DIST_EPSILON) / (t1 - t2);
 	}
 
-	t1 = DoublePrecisionDotProduct (trace->plane.normal, ctx->start) - trace->plane.dist;
-	t2 = DoublePrecisionDotProduct (trace->plane.normal, ctx->end) - trace->plane.dist;
-	midf = (t1 - DIST_EPSILON) / (t1 - t2);
+	t1 = DoublePrecisionDotProduct (trace->plane.normal, ctx->start) - (double)trace->plane.dist;
+	t2 = DoublePrecisionDotProduct (trace->plane.normal, ctx->end) - (double)trace->plane.dist;
+	midf = (t1 - (float)DIST_EPSILON) / (t1 - t2);
 
 	midf = CLAMP(0, midf, 1);
 	trace->fraction = midf;
@@ -1012,7 +1012,7 @@ void SV_ClipToLinks ( areanode_t *node, moveclip_t *clip )
 		trace.fraction < clip->trace.fraction)
 		{
 			trace.ent = touch;
-		 	if (clip->trace.startsolid)
+			if (clip->trace.startsolid)
 			{
 				clip->trace = trace;
 				clip->trace.startsolid = true;
